backend/requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
better-abc==0.0.3
black==21.7b0
bleach==3.3.0
CacheControl==0.12.10
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.4
Choco==1.0.5
click==8.0.1
colorama==0.4.4
cvxopt==1.2.6
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
firebase-admin==5.2.0
Flask==2.0.2
Flask-Cors==3.0.10
gcloud==0.18.3
google-api-core==1.31.1
google-api-python-client==2.34.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.44.0
google-cloud-vision==2.4.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.40.0
grpcio-tools==1.40.0
httplib2==0.20.2
idna==3.2
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jws==0.1.3
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.3
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==5.6.1
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
notebook-as-pdf==0.5.0
numpy==1.20.3
oauth2client==4.1.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.12.0
pyee==8.1.0
Pygments==2.9.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.5
Pyrebase4==4.5.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2021.1
pywin32==300
pywinpty==1.1.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
tomli==1.2.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
uritemplate==4.1.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==2.0.2
